% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCloudConvert.R
\name{convert_file}
\alias{convert_file}
\title{To convert file to other format}
\usage{
convert_file(api_key, input_format, output_format, input, file, dest_file)
}
\arguments{
\item{api_key}{CloudConvert API key}

\item{input_format}{The format of the input file}

\item{output_format}{The format the file should be converted to}

\item{input}{Set the method of uploading the input file}

\item{file}{Input file destination (including the name of file with extension)}

\item{dest_file}{Name where the downloaded file is saved (including the name of file with extension)}
}
\description{
To convert file to other format
}
\examples{
convert_file(api_key=api_key,input_format = "jpg",output_format = "pdf",input="upload",file="input file location",dest_file = "destination file location")
}

